Suggest applying the generated diffs
The output was expensive to calculate, and the diff output is machine generated.
`patch` is fully capable of applying it, so let people know that if they are
comfortable with diffs they are welcome to apply them.

Otherwise, people can still run the command to overwrite things if they have the
cycles.

Signed-off-by: Josh Soref <2119212+jsoref@users.noreply.github.com>