remove delete of extra CRDs dir
This was intended to help clean up the extra CRDs dir which was created
by mistake in #239.

However, a different makefile-modules PR (#240) meant that a manual
upgrade was required for downstream repos using makefile-modules anyway.

As such, this is just error-prone and liable to break ci (it should've
been `rm -rf` anyway).

Signed-off-by: Ashley Davis <ashley.davis@cyberark.com>